Consolidate Inspector state types/setup in reducer file
