+++ src/devtools/client/inspector/reducers/index.ts
@@ -5,7 +5,6 @@
 "use strict";
 
 import { InspectorAction } from "../actions";
-import { initialInspectorState, InspectorState } from "../state";
 const { prefs } = require("devtools/client/inspector/prefs");
 
 import boxModel from "devtools/client/inspector/boxmodel/reducers/box-model";
@@ -14,6 +13,18 @@ import markup from "devtools/client/inspector/markup/reducers/markup";
 import rules from "devtools/client/inspector/rules/reducers/rules";
 import computed from "devtools/client/inspector/computed/reducers";
 export { classList, markup, rules, computed, boxModel };
+
+export type InspectorActiveTab = "ruleview" | "layoutview" | "computedview" | "eventsview";
+
+export interface InspectorState {
+  activeTab: InspectorActiveTab;
+}
+
+export function initialInspectorState(): InspectorState {
+  return {
+    activeTab: prefs.activeTab,
+  };
+}
 
 export function inspector(
   state: InspectorState = initialInspectorState(),
