--- src/devtools/client/inspector/state/index.ts
@@ -1,13 +1,0 @@
-const { prefs } = require("devtools/client/inspector/prefs");
-
-export type InspectorActiveTab = "ruleview" | "layoutview" | "computedview" | "eventsview";
-
-export interface InspectorState {
-  activeTab: InspectorActiveTab;
-}
-
-export function initialInspectorState(): InspectorState {
-  return {
-    activeTab: prefs.activeTab,
-  };
-}
