+++ src/devtools/server/actors/utils/event-breakpoints.js
@@ -2,7 +2,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-"use strict";
+// Side-effectful import, needed to initialize prefs before read
+require("devtools/client/inspector/prefs");
 
 const Services = require("devtools/shared/services");
