+++ src/ui/state/index.ts
@@ -4,12 +4,11 @@ import { ASTState } from "devtools/client/debugger/src/reducers/ast";
 import type { AsyncRequestsState } from "devtools/client/debugger/src/reducers/async-requests";
 import { BoxModelState } from "devtools/client/inspector/boxmodel/reducers/box-model";
 import { BreakpointsState } from "devtools/client/debugger/src/selectors";
-import type { ChangesState } from "devtools/client/inspector/changes/reducers/changes";
 import type { EventListenersState } from "devtools/client/debugger/src/selectors";
 import type { FileSearchState } from "devtools/client/debugger/src/selectors";
 import { ContextMenusState } from "../reducers/contextMenus";
 import { ReactDevToolsState } from "./reactDevTools";
-import { InspectorState } from "devtools/client/inspector/state";
+import { InspectorState } from "devtools/client/inspector/reducers";
 import { MarkupState } from "devtools/client/inspector/markup/reducers/markup";
 import { ClassListState } from "devtools/client/inspector/rules/state/class-list";
 import { RulesState } from "devtools/client/inspector/rules/state/rules";
@@ -39,7 +38,6 @@ export interface UIState {
   asyncRequests: AsyncRequestsState;
   boxModel: BoxModelState;
   breakpoints: BreakpointsState;
-  changes: ChangesState;
   classList: ClassListState;
   computed: ComputedState;
   consoleUI: WebconsoleUIState;
